+++ src/index.ts
@@ -1,3 +1,3 @@
 export type Namespace = string | string[]
-export { Dispatch } from './buildDispatch'
+export { default as ConnectedProvider } from './ConnectedProvider'
 export { default as useConnectedReducer } from './useConnectedReducer'